requirements.txt
django
pandas
requests
openpyxl
gunicorn
python-dotenv
whitenoise
django-admin-soft-dashboard==1.0.11
scikit-learn
mysqlclient
pymysql
djangorestframework
django-cors-headers


# psycopg2-binary
# mysqlclient

#pk.eyJ1Ijoicm91dGUyNCIsImEiOiJjbHd5Z25oeWQxbDV5MnFxOHE4OGFla2o4In0.AWn6zJ26HiXyH04mIAq6Kg
